Extracted logic for creating execution strategy in separate helper class https://github.com/SoftUni-Internal/suls-issues/issues/3898
