+++ Open Judge System/Workers/OJS.Workers.LocalWorker/OJS.Workers.LocalWorker.csproj
@@ -80,6 +80,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Helpers\SubmissionJobHelper.cs" />
     <Compile Include="IJob.cs" />
     <Compile Include="LocalWorkerService.cs">
       <SubType>Component</SubType>
